requirements.txt
gradio==4.44.1
#openai==1.2.3

# Core LangChain Libraries
langchain>=0.3.0
langchain-core>=0.3.35
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# OpenAI Integration
langchain-openai>=0.3.5
openai>=1.62.0

# Tokenizer
tiktoken>=0.8.0

# Environment Management
python-dotenv>=1.0.0

# Utility Libraries
pydantic>=2.10.6
tenacity>=8.5.0

# HTTP Client
httpx>=0.28.1

# Others
orjson>=3.10.15
jiter>=0.8.2
regex>=2024.11.6
